# docs/architecture/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080/api
  description: Generated server url
...
paths:
...
  /pets/{id}:
    get:
      tags:
      - Pets
      summary: Get pet by ID
      description: Retrieves a pet by its ID
      operationId: getPetById
      parameters:
      - name: id
        in: path
        description: Pet ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Pet found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PetResponse"
        "404":
          description: Pet not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PetResponse"
    put:
      tags:
      - Pets
      summary: Update a pet
      description: Updates an existing pet
      operationId: updatePet
      parameters:
      - name: id
        in: path
        description: Pet ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePetRequest"
        required: true
      responses:
        "404":
          description: Pet not found
        "422":
          description: Validation error
        "200":
          description: Pet updated successfully
          headers:
            Location:
              description: URI of the updated pet
              style: simple
        "400":
          description: Invalid input
    delete:
      tags:
      - Pets
      summary: Delete a pet
      description: Deletes a pet by its ID
      operationId: deletePet
      parameters:
      - name: id
        in: path
        description: Pet ID
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Pet not found
        "204":
          description: Pet deleted successfully
...
  /pets:
    get:
      tags:
      - Pets
      summary: List all pets
      description: Retrieves a paginated list of pets with optional search
      operationId: listPets
      parameters:
      - name: page
        in: query
        description: Page number (0-based)
        required: false
        schema:
          type: integer
          format: int32
          default: 0
        example: 0
      - name: perPage
        in: query
        description: Items per page
        required: false
        schema:
          type: integer
          format: int32
          default: 10
        example: 10
      - name: search
        in: query
        description: Search terms
        required: false
        schema:
          type: string
          default: ""
        example: Fluffy
      - name: sort
        in: query
        description: Sort field
        required: false
        schema:
          type: string
          default: name
        example: name
      - name: direction
        in: query
        description: Sort direction (asc/desc)
        required: false
        schema:
          type: string
          default: asc
        example: asc
      responses:
        "200":
          description: Pets retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginationPetResponse"
    post:
      tags:
      - Pets
      summary: Create a new pet
      description: Creates a new pet in the system
      operationId: createPet
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePetRequest"
        required: true
      responses:
        "201":
          description: Pet created successfully
          headers:
            Location:
              description: URI of the created pet
              style: simple
        "422":
          description: Validation error
        "400":
          description: Invalid input
...
components:
  schemas:
...
    UpdatePetRequest:
      type: object
      properties:
        name:
          type: string
        species:
          type: string
        breed:
          type: string
        birth_date:
          type: string
          format: date
        notes:
          type: string
...
    CreatePetRequest:
      type: object
      properties:
        tutor_id:
          type: string
        name:
          type: string
        species:
          type: string
        breed:
          type: string
        birth_date:
          type: string
          format: date
        notes:
          type: string
...
    PaginationPetResponse:
      type: object
      properties:
        currentPage:
          type: integer
          format: int32
        perPage:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
        items:
          type: array
          items:
            $ref: "#/components/schemas/PetResponse"
    PetResponse:
      type: object
      properties:
        id:
          type: string
        tutor_id:
          type: string
        name:
          type: string
        species:
          type: string
        breed:
          type: string
        birth_date:
          type: string
          format: date
        notes:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...